moved logic for auto scroll to fieldwrapper
